+++ posthog/src/main/java/com/posthog/PostHogBeforeSend.kt
@@ -1,0 +1,5 @@
+package com.posthog
+
+public fun interface PostHogBeforeSend {
+    public fun run(event: PostHogEvent): PostHogEvent?
+}