+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next
 
+- feat: add support for beforeSend function to edit or drop events ([#266](https://github.com/PostHog/posthog-android/pull/266))
+    - Thanks @KopeikinaDarya 
+
 ## 3.19.2 - 2025-07-10
 
 - fix: enable gzip for /flags endpoint  ([#268](https://github.com/PostHog/posthog-android/pull/245)) and ([#268](https://github.com/PostHog/posthog-android/pull/264))